requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
mysqlclient==2.2.4
djangorestframework-simplejwt==5.3.1